fix negative reserve case for staking

+++ contracts/staking/IStaking.sol
@@ -19,7 +19,7 @@ interface IStaking is
     event LogStakingRewardsUpdated(address user, bytes32 targetId, bytes32 instanceId, uint256 componentId, uint256 bundleId, uint256 amount, uint256 newBalance);
     event LogStakingRewardsClaimed(address user, bytes32 targetId, bytes32 instanceId, uint256 componentId, uint256 bundleId, uint256 amount, uint256 newBalance);
 
-    event LogStakingDipBalanceChanged(uint256 stakeBalance, uint256 rewardBalance, uint256 actualBalance, uint reserves);
+    event LogStakingDipBalanceChanged(uint256 stakeBalance, uint256 rewardBalance, uint256 actualBalance, int256 reserves);
 
     function increaseRewardReserves(uint256 amount) external;
